Cleaned up the buffer helpers
